+++ kosinski.c
@@ -118,7 +118,7 @@ static void DoMatch(size_t distance, size_t length, size_t offset, void *user)
 	}
 }
 
-static unsigned int GetMatchCost(size_t distance, size_t length, void *user)
+static size_t GetMatchCost(size_t distance, size_t length, void *user)
 {
 	(void)user;
